Optimize progress view: remove 4 Shadow DOM components, add guard()
Remove LogEntry, TaskGroupItem, TaskGroupHeader, and LogPlaceholder
components (279 LOC) — each created a separate Shadow DOM boundary with
duplicated style sheets in the hot rendering path.

Inline their templates directly into TaskGroupList with Lit's guard()
directive to skip unchanged template creation. This eliminates N+2G
shadow roots (N messages + 2 per group) while preserving the same
change-detection behavior.

Also: inline shouldBeOpen() (single call site), increase markdown
cache from 500→2000, add .log-placeholder CSS class to layoutStyles.
