--- src/progressView/frontend/components/LogEntry.ts
@@ -1,42 +1,0 @@
-/**
- * Declarative log entry component.
- * Renders log messages using Lit templates directly in Shadow DOM.
- */
-
-// Third-party imports
-import { LitElement, html, nothing, type TemplateResult } from 'lit';
-import { customElement, property } from 'lit/decorators.js';
-
-// Local imports - formatters
-import { designTokens, commonViewStyles, codiconStyles } from '@shared/styles';
-import { formatLogEntry } from '../formatters';
-
-// Local imports - progress view styles
-import { logStyles } from '../styles/logStyles';
-
-// Local imports - shared schemas
-import type { LogMessageData } from '@shared/schemas';
-
-@customElement('log-entry')
-export class LogEntry extends LitElement {
-  static override styles = [
-    designTokens,
-    commonViewStyles,
-    codiconStyles,
-    ...logStyles,
-  ];
-
-  @property({ attribute: false }) message!: LogMessageData;
-  @property({ attribute: false }) defaultOpen = false;
-
-  override render(): TemplateResult | typeof nothing {
-    if (!this.message) return nothing;
-    return formatLogEntry(this.message, { defaultOpen: this.defaultOpen });
-  }
-}
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'log-entry': LogEntry;
-  }
-}
