+++ src/progressView/frontend/formatters/baseLogFormatter.ts
@@ -26,18 +26,3 @@ export function safeFormat<T>(
     return { ok: false, error: errorMsg };
   }
 }
-
-/**
- * Determine whether a details element should be open.
- * Returns undefined when no preference is set, allowing formatters to use their own defaults.
- */
-export function shouldBeOpen(
-  messageType: string,
-  options: FormatOptions | undefined,
-  autoExpandedTypes: Set<string>,
-): boolean | undefined {
-  if (!options) return undefined;
-  if (options.preservedOpen !== undefined) return options.preservedOpen;
-  if (options.defaultOpen && autoExpandedTypes.has(messageType)) return true;
-  return undefined;
-}
